many things (see desc)
changed naming such that this.room.players.players and similar setups for the other objects are now this.room.players.objects

changed id system so that every single object, be it a shot or an enemy or a player or a pickup have unique keys so as long as you retain the keys you should be able to tell objects apart and not have to go find them somewhere in the larger objects

except for the players (who already are being constructed by the Socket constructor), all objects now have actual object constructors, so you can figure out the type by just asking for the name of the constructor. This means we dont need the obj_type attribute anymore
(ie. obj.obj_type == "enemy"  is now just obj.constructor.name == "Enemy")